Update to convert csv data on the fly
