+++ src/app.py
@@ -1,6 +1,7 @@
 import streamlit as st
 import pandas as pd
 import os
+import helpers.converters as conv
 
 
 def GetBenchmarkData() -> pd.DataFrame:
@@ -8,10 +9,9 @@ def GetBenchmarkData() -> pd.DataFrame:
   # We will likely change this in the future to pull from
   # a URL or dataabase
   flatDataPath = os.path.join(os.path.dirname(__file__), '..', 'data', 'benchmark-data.csv')
-  typeDataPath = os.path.join(os.path.dirname(__file__), '..', 'data', 'benchmark-data.json')
 
   flatData = pd.read_csv(flatDataPath)
-  typeData = pd.read_json(typeDataPath)
+  typeData = conv.ConvertFlatToType(flatData=flatData)
   return flatData, typeData
